+++ src/main/java/com/ride/domain/rider/Rider.java
@@ -1,12 +1,12 @@
 package com.ride.domain.rider;
 
 public class Rider {
-    private int riderId;
+    private Long riderId;
     private String riderName;
     private String phone;
     private String email;
 
-    public Rider(int riderId, String riderName, String phone, String email) {
+    public Rider(Long riderId, String riderName, String phone, String email) {
         this.riderId = riderId;
         this.riderName = riderName;
         this.phone = phone;
@@ -14,7 +14,7 @@ public class Rider {
 
     }
 
-    public int getRiderId() { return riderId; }
+    public Long getRiderId() { return riderId; }
     public String getRiderName() { return riderName; }
     public String getPhone() { return phone; }
     public String getEmail() { return email; }
