+++ src/main/java/com/ride/dto/request/CreateRideRequestDTO.java
@@ -3,7 +3,7 @@ package com.ride.dto.request;
 import com.ride.domain.ride.Location;
 
 public class CreateRideRequestDTO {
-    public int riderId;
+    public Long riderId;
     public String pickup;
     public String dropoff;
     public Location pickupLocation;
